+++ src/test/java/org/cactoos/list/SumOfIntTest.java
@@ -23,49 +23,45 @@
  */
 package org.cactoos.list;
 
-import java.util.Iterator;
 import org.cactoos.Scalar;
+import org.hamcrest.MatcherAssert;
+import org.hamcrest.Matchers;
+import org.junit.Test;
 
 /**
- * Real total of numbers.
- *
- * <p>There is no thread-safety guarantee.
+ * Test case for {@link SumOfInt}.
  *
  * @author Vseslav Sekorin (vssekorin@gmail.com)
  * @version $Id$
  * @since 0.9
+ * @checkstyle JavadocMethodCheck (500 lines)
+ * @checkstyle MagicNumberCheck (500 lines)
  */
-public final class RealTotal implements Scalar<Double> {
+public final class SumOfIntTest {
 
-    /**
-     * The iterable.
-     */
-    private final Iterable<Scalar<Number>> src;
-
-    /**
-     * Ctor.
-     * @param src The iterable
-     */
-    public RealTotal(final Iterable<Scalar<Number>> src) {
-        this.src = src;
+    @Test
+    public void withVarargsCtor() throws Exception {
+        MatcherAssert.assertThat(
+            new SumOfInt(
+                () -> 1,
+                () -> 2,
+                () -> 3
+            ).value(),
+            Matchers.equalTo(6L)
+        );
     }
 
-    /**
-     * Ctor.
-     * @param src Numbers
-     */
-    @SafeVarargs
-    public RealTotal(final Scalar<Number>... src) {
-        this(new ArrayAsIterable<>(src));
-    }
-
-    @Override
-    public Double value() throws Exception {
-        final Iterator<Scalar<Number>> numbers = this.src.iterator();
-        Double result =  0.;
-        while (numbers.hasNext()) {
-            result += numbers.next().value().doubleValue();
-        }
-        return result;
+    @Test
+    public void withIterCtor() throws Exception {
+        MatcherAssert.assertThat(
+            new SumOfInt(
+                new ArrayAsIterable<Scalar<Number>>(
+                    () -> 7,
+                    () -> 8,
+                    () -> 10
+                )
+            ).value(),
+            Matchers.equalTo(25L)
+        );
     }
 }
